airflow-zrch-docker/requirements.txt
pandas==2.2.2
apache-airflow-providers-google==8.10.0
SQLAlchemy==1.4.44
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.8
sqlparse==0.4.4
openpyxl==3.1.4

